+++ translations/nl.php
@@ -1,0 +1,51 @@
+<?php
+
+/**
+ * Dutch translations for SearchProtect module.
+ *
+ * @author    Tecnoacquisti.com
+ * @copyright 2026 Tecnoacquisti.com - Arte e Informatica di Loris Modena e c. s.a.s.
+ * @license   MIT
+ */
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+global $_MODULE;
+
+$_MODULE = [];
+$_MODULE['<{searchprotect}prestashop>searchprotect_fa2b150d92346f2ac2273a3034e10ce0'] = 'Search Protect';
+$_MODULE['<{searchprotect}prestashop>searchprotect_fade827e0c373c437207ee6268227ead'] = 'Blokkeert misvormde of te lange zoekopdrachten om DoS-aanvallen op MariaDB te voorkomen.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_bf19c33ddba7844c76a4aedeb64668dd'] = 'De maximale lengte van de zoekopdracht moet tussen 10 en 2000 liggen.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_de506dcedbe68a9fd6925f10646d6594'] = 'Het maximale aantal & moet tussen 1 en 100 liggen.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_2879ba8c8ea279d9b0fc0cd484fe74f1'] = 'Het maximale aantal ?page= moet tussen 1 en 50 liggen.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ffd00d70282cd23caa37e19ed230b339'] = 'De blokkeringsduur moet tussen 60 en 86400 seconden liggen.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_9a7e58ee13cb727cb7c4acc68f4cafe5'] = 'Instellingen opgeslagen.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_f52299428e080343f571f7795bbeb884'] = 'Beschermingsregels';
+$_MODULE['<{searchprotect}prestashop>searchprotect_3af74250eda4f2cf8fc756842a897136'] = 'Maximale zoeklengte (tekens)';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c04ff056933f62b62ade8a8a6814bbaf'] = 'Zoekopdrachten langer dan deze waarde worden geblokkeerd. Aanbevolen: 100.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_3fb9fceb0a504891cf29b7965cfa4728'] = 'Maximale &-voorkomens';
+$_MODULE['<{searchprotect}prestashop>searchprotect_d4252ec713a1efaafb7e736f990fdce3'] = 'Blokkeert zoekopdrachten met herhaalde &-codering (DoS-patroon). Aanbevolen: 5.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_163a953f8d4a8110eb0959fb2825b0a1'] = 'Maximale ?page=-voorkomens';
+$_MODULE['<{searchprotect}prestashop>searchprotect_89c6f1ed277c210229f84b02e9422356'] = 'Blokkeert herhaalde ?page=-injectie. Aanbevolen: 3.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_677a741921fcb9cafcd1b516f643a64f'] = 'IP-blokkeringsduur (seconden)';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ff3e8abd787060e19248da933d51d889'] = 'Hoe lang een overtredend IP-adres wordt geblokkeerd. Standaard: 3600 (1 uur).';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c218ee1f0300a1c69a0f97d6d2872fde'] = 'Logboek inschakelen';
+$_MODULE['<{searchprotect}prestashop>searchprotect_93cba07454f06a4a960172bbd6e2a435'] = 'Ja';
+$_MODULE['<{searchprotect}prestashop>searchprotect_bafd7322c6e97d25b6299b5d6fe8920b'] = 'Nee';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c9cc8cce247e49bae79f15173ce97354'] = 'Opslaan';
+$_MODULE['<{searchprotect}prestashop>searchprotect_1989b20b0791aa82dc5e0d6ce60f400d'] = 'Laatste 100 geblokkeerde verzoeken';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ac9e63f88dcb464b7fce17f1745043ef'] = 'Nog geen aanvallen geregistreerd.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_44749712dbec183e983dcd78a7736c41'] = 'Datum';
+$_MODULE['<{searchprotect}prestashop>searchprotect_a12a3079e14ced46e69ba52b8a90b21a'] = 'IP';
+$_MODULE['<{searchprotect}prestashop>searchprotect_fffcb4d1e12c96b2f571f77ad25878b8'] = 'Reden';
+$_MODULE['<{searchprotect}prestashop>searchprotect_cf0f75226d64edb0e6f8d6546103aa0f'] = 'Zoekopdracht (afgekapt)';
+$_MODULE['<{searchprotect}prestashop>copyright_948a2e3548aaf7f9941a3192fa607d51'] = 'Credits';
+$_MODULE['<{searchprotect}prestashop>copyright_f32af624cf9b69880166c492757ea91c'] = 'PrestaShop-modules';
+$_MODULE['<{searchprotect}prestashop>copyright_bb0753bfcd40d437efac5cad0b16eee4'] = 'Wij bieden maatwerkoplossingen om uw online bedrijf te helpen ontwikkelen';
+$_MODULE['<{searchprotect}prestashop>copyright_2434b895cd68fd06b5db048941cc503a'] = 'Tecnoacquisti.com® helpt bedrijven hun producten online te verkopen met moderne en betrouwbare systemen.';
+$_MODULE['<{searchprotect}prestashop>copyright_4c4e11b55bcd88ee46b8196d78f6e023'] = 'Wij zijn PrestaShop-partner en ontwikkelen modules, diensten en oplossingen voor al uw behoeften.';
+$_MODULE['<{searchprotect}prestashop>copyright_1741ec6ed69cf571adac9eeb641c4f87'] = 'Ondersteuning';
+$_MODULE['<{searchprotect}prestashop>copyright_04a14c4bfab97f16089cde138d3b9476'] = 'Check-up';
+$_MODULE['<{searchprotect}prestashop>copyright_affa9089ac0a7a008631522d445df967'] = 'PrestaShop-hosting';
+$_MODULE['<{searchprotect}prestashop>copyright_fd22b349b10c8a5c59ec7d07450e6840'] = 'Tecnoacquisti.com® is een geregistreerd handelsmerk van Arte e Informatica';