+++ translations/pl.php
@@ -1,0 +1,51 @@
+<?php
+
+/**
+ * Polish translations for SearchProtect module.
+ *
+ * @author    Tecnoacquisti.com
+ * @copyright 2026 Tecnoacquisti.com - Arte e Informatica di Loris Modena e c. s.a.s.
+ * @license   MIT
+ */
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+global $_MODULE;
+
+$_MODULE = [];
+$_MODULE['<{searchprotect}prestashop>searchprotect_fa2b150d92346f2ac2273a3034e10ce0'] = 'Search Protect';
+$_MODULE['<{searchprotect}prestashop>searchprotect_fade827e0c373c437207ee6268227ead'] = 'Blokuje nieprawidłowe lub zbyt długie zapytania wyszukiwania, aby zapobiegać atakom DoS na MariaDB.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_bf19c33ddba7844c76a4aedeb64668dd'] = 'Maksymalna długość zapytania musi wynosić od 10 do 2000.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_de506dcedbe68a9fd6925f10646d6594'] = 'Maksymalna liczba znaków & musi wynosić od 1 do 100.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_2879ba8c8ea279d9b0fc0cd484fe74f1'] = 'Maksymalna liczba wystąpień ?page= musi wynosić od 1 do 50.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ffd00d70282cd23caa37e19ed230b339'] = 'Czas blokady musi wynosić od 60 do 86400 sekund.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_9a7e58ee13cb727cb7c4acc68f4cafe5'] = 'Ustawienia zapisane.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_f52299428e080343f571f7795bbeb884'] = 'Reguły ochrony';
+$_MODULE['<{searchprotect}prestashop>searchprotect_3af74250eda4f2cf8fc756842a897136'] = 'Maksymalna długość zapytania (znaki)';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c04ff056933f62b62ade8a8a6814bbaf'] = 'Zapytania dłuższe niż ta wartość zostaną zablokowane. Zalecane: 100.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_3fb9fceb0a504891cf29b7965cfa4728'] = 'Maksymalna liczba wystąpień &';
+$_MODULE['<{searchprotect}prestashop>searchprotect_d4252ec713a1efaafb7e736f990fdce3'] = 'Blokuje zapytania z powtarzaną enkodowaną wartością & (wzorzec DoS). Zalecane: 5.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_163a953f8d4a8110eb0959fb2825b0a1'] = 'Maksymalna liczba wystąpień ?page=';
+$_MODULE['<{searchprotect}prestashop>searchprotect_89c6f1ed277c210229f84b02e9422356'] = 'Blokuje powtarzające się wstrzyknięcia ?page=. Zalecane: 3.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_677a741921fcb9cafcd1b516f643a64f'] = 'Czas blokady IP (sekundy)';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ff3e8abd787060e19248da933d51d889'] = 'Jak długo blokować naruszający adres IP. Domyślnie: 3600 (1 godzina).';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c218ee1f0300a1c69a0f97d6d2872fde'] = 'Włącz rejestrowanie';
+$_MODULE['<{searchprotect}prestashop>searchprotect_93cba07454f06a4a960172bbd6e2a435'] = 'Tak';
+$_MODULE['<{searchprotect}prestashop>searchprotect_bafd7322c6e97d25b6299b5d6fe8920b'] = 'Nie';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c9cc8cce247e49bae79f15173ce97354'] = 'Zapisz';
+$_MODULE['<{searchprotect}prestashop>searchprotect_1989b20b0791aa82dc5e0d6ce60f400d'] = 'Ostatnie 100 zablokowanych żądań';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ac9e63f88dcb464b7fce17f1745043ef'] = 'Nie zarejestrowano jeszcze żadnych ataków.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_44749712dbec183e983dcd78a7736c41'] = 'Data';
+$_MODULE['<{searchprotect}prestashop>searchprotect_a12a3079e14ced46e69ba52b8a90b21a'] = 'IP';
+$_MODULE['<{searchprotect}prestashop>searchprotect_fffcb4d1e12c96b2f571f77ad25878b8'] = 'Powód';
+$_MODULE['<{searchprotect}prestashop>searchprotect_cf0f75226d64edb0e6f8d6546103aa0f'] = 'Zapytanie (skrócone)';
+$_MODULE['<{searchprotect}prestashop>copyright_948a2e3548aaf7f9941a3192fa607d51'] = 'Autorzy';
+$_MODULE['<{searchprotect}prestashop>copyright_f32af624cf9b69880166c492757ea91c'] = 'Moduły PrestaShop';
+$_MODULE['<{searchprotect}prestashop>copyright_bb0753bfcd40d437efac5cad0b16eee4'] = 'Dostarczamy rozwiązania dostosowane do potrzeb, aby pomóc rozwijać Twój biznes online';
+$_MODULE['<{searchprotect}prestashop>copyright_2434b895cd68fd06b5db048941cc503a'] = 'Tecnoacquisti.com® pomaga firmom sprzedawać produkty online za pomocą nowoczesnych i niezawodnych systemów.';
+$_MODULE['<{searchprotect}prestashop>copyright_4c4e11b55bcd88ee46b8196d78f6e023'] = 'Jesteśmy partnerem PrestaShop i tworzymy moduły, usługi oraz rozwiązania dla wszystkich Twoich potrzeb.';
+$_MODULE['<{searchprotect}prestashop>copyright_1741ec6ed69cf571adac9eeb641c4f87'] = 'Pomoc';
+$_MODULE['<{searchprotect}prestashop>copyright_04a14c4bfab97f16089cde138d3b9476'] = 'Check-up';
+$_MODULE['<{searchprotect}prestashop>copyright_affa9089ac0a7a008631522d445df967'] = 'Hosting PrestaShop';
+$_MODULE['<{searchprotect}prestashop>copyright_fd22b349b10c8a5c59ec7d07450e6840'] = 'Tecnoacquisti.com® jest zarejestrowanym znakiem towarowym Arte e Informatica';