+++ translations/ro.php
@@ -1,0 +1,51 @@
+<?php
+
+/**
+ * Romanian translations for SearchProtect module.
+ *
+ * @author    Tecnoacquisti.com
+ * @copyright 2026 Tecnoacquisti.com - Arte e Informatica di Loris Modena e c. s.a.s.
+ * @license   MIT
+ */
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+global $_MODULE;
+
+$_MODULE = [];
+$_MODULE['<{searchprotect}prestashop>searchprotect_fa2b150d92346f2ac2273a3034e10ce0'] = 'Search Protect';
+$_MODULE['<{searchprotect}prestashop>searchprotect_fade827e0c373c437207ee6268227ead'] = 'Blochează interogările de căutare malformate sau prea lungi pentru a preveni atacurile DoS asupra MariaDB.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_bf19c33ddba7844c76a4aedeb64668dd'] = 'Lungimea maximă a interogării trebuie să fie între 10 și 2000.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_de506dcedbe68a9fd6925f10646d6594'] = 'Numărul maxim de & trebuie să fie între 1 și 100.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_2879ba8c8ea279d9b0fc0cd484fe74f1'] = 'Numărul maxim de ?page= trebuie să fie între 1 și 50.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ffd00d70282cd23caa37e19ed230b339'] = 'Durata blocării trebuie să fie între 60 și 86400 de secunde.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_9a7e58ee13cb727cb7c4acc68f4cafe5'] = 'Setări salvate.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_f52299428e080343f571f7795bbeb884'] = 'Reguli de protecție';
+$_MODULE['<{searchprotect}prestashop>searchprotect_3af74250eda4f2cf8fc756842a897136'] = 'Lungime maximă interogare (caractere)';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c04ff056933f62b62ade8a8a6814bbaf'] = 'Interogările mai lungi decât această valoare vor fi blocate. Recomandat: 100.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_3fb9fceb0a504891cf29b7965cfa4728'] = 'Apariții maxime ale &';
+$_MODULE['<{searchprotect}prestashop>searchprotect_d4252ec713a1efaafb7e736f990fdce3'] = 'Blochează interogările cu codificare & repetată (model DoS). Recomandat: 5.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_163a953f8d4a8110eb0959fb2825b0a1'] = 'Apariții maxime ale ?page=';
+$_MODULE['<{searchprotect}prestashop>searchprotect_89c6f1ed277c210229f84b02e9422356'] = 'Blochează injectările repetate ?page=. Recomandat: 3.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_677a741921fcb9cafcd1b516f643a64f'] = 'Durata blocării IP (secunde)';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ff3e8abd787060e19248da933d51d889'] = 'Cât timp se blochează un IP ofensator. Implicit: 3600 (1 oră).';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c218ee1f0300a1c69a0f97d6d2872fde'] = 'Activează jurnalizarea';
+$_MODULE['<{searchprotect}prestashop>searchprotect_93cba07454f06a4a960172bbd6e2a435'] = 'Da';
+$_MODULE['<{searchprotect}prestashop>searchprotect_bafd7322c6e97d25b6299b5d6fe8920b'] = 'Nu';
+$_MODULE['<{searchprotect}prestashop>searchprotect_c9cc8cce247e49bae79f15173ce97354'] = 'Salvează';
+$_MODULE['<{searchprotect}prestashop>searchprotect_1989b20b0791aa82dc5e0d6ce60f400d'] = 'Ultimele 100 de cereri blocate';
+$_MODULE['<{searchprotect}prestashop>searchprotect_ac9e63f88dcb464b7fce17f1745043ef'] = 'Nu există încă atacuri înregistrate.';
+$_MODULE['<{searchprotect}prestashop>searchprotect_44749712dbec183e983dcd78a7736c41'] = 'Dată';
+$_MODULE['<{searchprotect}prestashop>searchprotect_a12a3079e14ced46e69ba52b8a90b21a'] = 'IP';
+$_MODULE['<{searchprotect}prestashop>searchprotect_fffcb4d1e12c96b2f571f77ad25878b8'] = 'Motiv';
+$_MODULE['<{searchprotect}prestashop>searchprotect_cf0f75226d64edb0e6f8d6546103aa0f'] = 'Interogare (trunchiată)';
+$_MODULE['<{searchprotect}prestashop>copyright_948a2e3548aaf7f9941a3192fa607d51'] = 'Credite';
+$_MODULE['<{searchprotect}prestashop>copyright_f32af624cf9b69880166c492757ea91c'] = 'Module PrestaShop';
+$_MODULE['<{searchprotect}prestashop>copyright_bb0753bfcd40d437efac5cad0b16eee4'] = 'Oferim soluții personalizate pentru a vă ajuta să vă dezvoltați afacerea online';
+$_MODULE['<{searchprotect}prestashop>copyright_2434b895cd68fd06b5db048941cc503a'] = 'Tecnoacquisti.com® ajută companiile să își vândă produsele online cu sisteme moderne și fiabile.';
+$_MODULE['<{searchprotect}prestashop>copyright_4c4e11b55bcd88ee46b8196d78f6e023'] = 'Suntem partener PrestaShop și dezvoltăm module, servicii și soluții pentru toate nevoile dumneavoastră.';
+$_MODULE['<{searchprotect}prestashop>copyright_1741ec6ed69cf571adac9eeb641c4f87'] = 'Asistență';
+$_MODULE['<{searchprotect}prestashop>copyright_04a14c4bfab97f16089cde138d3b9476'] = 'Check-up';
+$_MODULE['<{searchprotect}prestashop>copyright_affa9089ac0a7a008631522d445df967'] = 'Găzduire PrestaShop';
+$_MODULE['<{searchprotect}prestashop>copyright_fd22b349b10c8a5c59ec7d07450e6840'] = 'Tecnoacquisti.com® este o marcă înregistrată a Arte e Informatica';