+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [1.0.2] - 2026-06-26
+
+### Added
+
+- Added English, Italian, Spanish, French, German, Portuguese, Polish, Romanian, and Dutch translations.
+- Updated module logo asset.
+
+### Fixed
+
+- Fixed module file comment spacing for PrestaShop license validation.
+
 ## [1.0.1] - 2026-06-26
 
 ### Fixed
@@ -38,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-<!-- [Unreleased]: https://github.com/ArteInfoRM/searchprotect/compare/v1.0.1...HEAD -->
+<!-- [Unreleased]: https://github.com/ArteInfoRM/searchprotect/compare/v1.0.2...HEAD -->
+<!-- [1.0.2]: https://github.com/ArteInfoRM/searchprotect/compare/v1.0.1...v1.0.2 -->
 <!-- [1.0.1]: https://github.com/ArteInfoRM/searchprotect/compare/v1.0.0...v1.0.1 -->
 <!-- [1.0.0]: https://github.com/ArteInfoRM/searchprotect/releases/tag/v1.0.0 -->